clean tensor.h
