+++ paddle/fluid/operators/math/cross_entropy.h
@@ -17,6 +17,7 @@ limitations under the License. */
 
 #include "paddle/fluid/framework/eigen.h"
 #include "paddle/fluid/framework/tensor.h"
+#include "paddle/fluid/platform/float16.h"
 #include "paddle/phi/core/hostdevice.h"
 
 namespace paddle {
