+++ paddle/phi/core/dense_tensor.h
@@ -22,7 +22,7 @@ limitations under the License. */
 /* @jim19930609: Move to MKLDNN_Tensor in the future
  */
 #ifdef PADDLE_WITH_MKLDNN
-#include "dnnl.hpp"
+#include "dnnl.hpp"  // NOLINT
 #endif
 
 namespace phi {
@@ -174,6 +174,36 @@ class DenseTensor : public TensorBase,
   /* Temporarily put InplaceVersion inside DenseTensor.
   Will move to AutogradMeta as soon as we switch to Eager Dygraph.
   */
+  /*
+  NOTE(liym27): [ What is TensorInplaceVersion used for? ]
+
+  TensorInplaceVersion is a version counter and every Tensor has a version
+  counter. It's used to check whether an inplace operation will result in an
+  incorrect gradient calculation. Version is incremented when the data of the
+  Variable is modified in place.
+
+  - Question: In what scenarios will version counters be shared?
+  - Answer: When two Variables/VarBases share the same C++ Tensor(its Allocation
+  may change), both of them share the same version counter. For examples:
+   1. `z = paddle.assign(input=x, output=y)`, `z` shares the same version
+  counter of `y` because z and y is the same VarBase;
+   2. `y = x.detach()`, `y` shares the same version counter of `x`.
+
+  - Question: In what scenarios will version counters NOT be shared?
+  - Answer: Replacing a `Variable`'s data by calling
+  `Tensor::ShareDataWith(...)` or `Tensor::ShareBufferWith(...)`. Because they
+  share the same Allocation but not framework::Tensor.
+
+  - Question: Why put the inplace_version_counter_ in framework::Tensor instead
+  of Allocation or Variable?
+  - Answer:
+   1. Tensor can call ResetHolder() to reset the corresponding Allocation so
+  that the inplace_version_counter_ changes if it's in Allocation, which will
+  lead to confusing information about inplace version.
+   2. If inplace_version_counter_ is in Variable, different VariableWrappers
+   should be able to share the same Variable. However, a VariableWrapper hold a
+   Variable object but not a pointer.
+ */
   class InplaceVersion {
    public:
     bool IsUnique() const { return inplace_version_ == 0; }
